requirements.txt
google-generativeai
pyyaml
nbformat
papermill
pymupdf
langgraph
pandas
numpy
scikit-learn
